+++ explcheck/testfiles/s413-03.lua
@@ -9,7 +9,7 @@ local filename = "s413-03.tex"
 local file = assert(io.open(filename, "r"))
 local content = assert(file:read("*a"))
 assert(file:close())
-local options = {expl3_detection_strategy = "always"}
+local options = {expl3_detection_strategy = "always", ignored_issues = {"w416"}}
 local issues = new_issues(filename, options)
 local results = {}