+++ explcheck/testfiles/w419-04.lua
@@ -5,12 +5,12 @@ local lexical_analysis = require("explcheck-lexical-analysis")
 local syntactic_analysis = require("explcheck-syntactic-analysis")
 local semantic_analysis = require("explcheck-semantic-analysis")
 
-local filename = "e408-02.tex"
+local filename = "w419-04.tex"
 
 local file = assert(io.open(filename, "r"))
 local content = assert(file:read("*a"))
 assert(file:close())
-local options = {expl3_detection_strategy = "always", ignored_issues = {"w416"}}
+local options = {expl3_detection_strategy = "always", ignored_issues = {"s413"}}
 local issues = new_issues(filename, options)
 local results = {}
 
@@ -19,14 +19,14 @@ lexical_analysis.process(filename, content, issues, results, options)
 syntactic_analysis.process(filename, content, issues, results, options)
 semantic_analysis.process(filename, content, issues, results, options)
 
-assert(#issues.errors == 1)
-assert(#issues.warnings == 0)
+assert(#issues.errors == 0)
+assert(#issues.warnings == 1)
 
-local expected_line_numbers = {{7, 8}}
-for index, err in ipairs(issues.sort(issues.errors)) do
-  assert(err[1] == "e408")
-  assert(err[2] == "calling an undefined function")
-  local byte_range = err[3]
+local expected_line_numbers = {{1, 2}}
+for index, warning in ipairs(issues.sort(issues.warnings)) do
+  assert(warning[1] == "w419")
+  assert(warning[2] == "using an undeclared variable or constant")
+  local byte_range = warning[3]
   local start_line_number = utils.convert_byte_to_line_and_column(results.line_starting_byte_numbers, byte_range:start())
   local end_line_number = utils.convert_byte_to_line_and_column(results.line_starting_byte_numbers, byte_range:stop())
   assert(start_line_number == expected_line_numbers[index][1])